Consolidate WP Codebox contract adapter
