+++ wordpress/index.js
@@ -38,15 +38,7 @@ module.exports = {
 	...require('./lib/audit-fanout-runtime-provider'),
 	...require('./lib/agent-terminal-actions'),
 	...require('./lib/agent-task-runner-spec'),
-	...require('./lib/wp-codebox-apply-adapter'),
-	...require('./lib/codebox-client'),
-	...require('./lib/codebox-provider-adapter'),
-	...require('./lib/wp-codebox-recipe-helper'),
-	...require('./lib/wp-codebox-resolver'),
-	...require('./lib/wp-codebox-fuzz-plan'),
-	...require('./lib/wp-codebox-artifacts'),
-	...require('./lib/wp-codebox-browser-coverage'),
-	...require('./lib/wp-codebox-fuzz-run'),
+	...require('./lib/datamachine-agent-ci-plan'),
 	...require('./lib/wordpress-workload-profile'),
 	...require('./lib/wordpress-workload-scale-profile'),
 	...require('./lib/wordpress-fuzz-manifest'),
